ImapProtocol: Add STATUS command support. (#424)
This library currently contains some functions
that purport to get the status of a folder, but
these are misleading because they issue the
EXAMINE command, not the STATUS command. This library
actually currently lacks support for the STATUS command,
and this command may be necessary to get some information
that the EXAMINE command will not provide.

This adds support for STATUS at the protocol level
and the folder level, and clarifies the difference
in the documentation.